+++ VFS/include/VFS/VFSSeqToRandomWrapper.h
@@ -9,8 +9,8 @@ public:
     VFSSeqToRandomROWrapperFile(const VFSFilePtr &_file_to_wrap);
     ~VFSSeqToRandomROWrapperFile();
     
-    virtual int Open(int _flags, const VFSCancelChecker &_cancel_checker) override;
-    int Open(int _flags,
+    virtual int Open(unsigned long _flags, const VFSCancelChecker &_cancel_checker) override;
+    int Open(unsigned long _flags,
              const VFSCancelChecker &_cancel_checker,
              function<void(uint64_t _bytes_proc, uint64_t _bytes_total)> _progress);
     virtual int Close() override;
@@ -40,7 +40,7 @@ private:
     };
     
     VFSSeqToRandomROWrapperFile(const char* _relative_path, const VFSHostPtr &_host, shared_ptr<Backend> _backend);
-    int OpenBackend(int _flags,
+    int OpenBackend(unsigned long _flags,
                     VFSCancelChecker _cancel_checker,
                     function<void(uint64_t _bytes_proc, uint64_t _bytes_total)> _progress);
