+++ VFS/source/ArcLA/File.h
@@ -16,7 +16,7 @@ public:
     ~File();
     
     
-    virtual int     Open(int _open_flags, const VFSCancelChecker &_cancel_checker) override;
+    virtual int     Open(unsigned long _open_flags, const VFSCancelChecker &_cancel_checker) override;
     virtual bool    IsOpened() const override;
     virtual int     Close() override;
     virtual ssize_t Read(void *_buf, size_t _size) override;
