+++ VFS/source/ArcUnRAR/File.cpp
@@ -19,7 +19,7 @@ File::~File()
     assert(m_ExtractionRunning == false);
 }
 
-int File::Open(int _open_flags, const VFSCancelChecker &_cancel_checker)
+int File::Open(unsigned long _open_flags, const VFSCancelChecker &_cancel_checker)
 {
     if( strlen(Path()) < 2 || Path()[0] != '/' )
         return SetLastError(VFSError::NotFound);
