+++ VFS/source/Native/File.cpp
@@ -20,7 +20,7 @@ File::~File()
     Close();
 }
 
-int File::Open(int _open_flags, const VFSCancelChecker &_cancel_checker)
+int File::Open(unsigned long _open_flags, const VFSCancelChecker &_cancel_checker)
 {
     auto &io = RoutedIO::Default;
     auto fs_info = NativeFSManager::Instance().VolumeFromPath(Path());
