+++ VFS/source/NetSFTP/File.cpp
@@ -16,7 +16,7 @@ File::~File()
     Close();
 }
 
-int File::Open(int _open_flags, const VFSCancelChecker &_cancel_checker)
+int File::Open(unsigned long _open_flags, const VFSCancelChecker &_cancel_checker)
 {
     if(IsOpened())
         Close();
